Collapse IsomorphicFixieClient and FixieClient, move to parameters-as-objects style
