+++ packages/fixie/index.ts
@@ -1,3 +1,2 @@
 export { FixieClient } from './src/client.js';
-export { IsomorphicFixieClient } from './src/isomorphic-client.js';
 export * from './src/sidekick-types.js';
