+++ packages/fixie/web.ts
@@ -1,4 +1,4 @@
-export { IsomorphicFixieClient } from './src/isomorphic-client.js';
+export { FixieClient } from './src/client.js';
 export {
   InlineFixieEmbed,
   FloatingFixieEmbed,
